[feat] 매칭확인뷰에서 받은 요청 migration
